requirements.txt
python-telegram-bot==20.3
transformers==4.41.2
torch==2.3.0+cu118 --index-url https://download.pytorch.org/whl/cu118
langchain-community==0.2.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
gitpython==3.1.37
accelerate==0.30.1
bitsandbytes==0.43.0
huggingface-hub==0.19.4
chardet==5.2.0
python-slugify==8.0.4